Move transports to example directory
