+++ example/teamclient/main.go
@@ -3,10 +3,11 @@ package main
 import (
 	"log"
 
+	"github.com/rsteube/carapace"
+
 	"github.com/reeflective/team/client"
 	"github.com/reeflective/team/client/commands"
-	grpc "github.com/reeflective/team/transports/grpc/client"
-	"github.com/rsteube/carapace"
+	grpc "github.com/reeflective/team/example/transports/grpc/client"
 )
 
 // main shows how to use a remote teamclient with a gRPC backend (transport & RPC).
